Remove user test

+++ data/backend/src/gateway/gateway.ts
@@ -114,7 +114,9 @@ export class Gateway
       user.uid > targetUser.uid
         ? user.uid + '' + targetUser.uid
         : targetUser.uid + '' + user.uid;
-    this.server.emit(payload.target + '⌲', {
+    client.join(listener + '⌲');
+    targetStatus.client.join(listener + '⌲');
+    this.server.to(payload.target + '⌲').emit(payload.target + '⌲', {
       source: payload.source,
       target: payload.target,
       listener: listener + '⌲',
@@ -153,7 +155,7 @@ export class Gateway
     else if (cUser.isMuted) throw new WsException("You're muted in that room");
     let role = await this.roomService.getRole(room, user.uid);
     if (!role) role = userRole.none;
-    this.server.emit(payload.target, {
+    this.server.to(payload.target).emit(payload.target, {
       source: payload.source,
       target: payload.target,
       action: actionTypes.msg,
@@ -185,7 +187,7 @@ export class Gateway
     if (!user) throw new WsException(payload.source + ' not found');
     if (await this.roomService.isUserInRoom(payload.room, user.uid))
       throw new WsException("You're already in that room");
-    const roomsJoined = await this.roomService.findAllRoomUser(payload.source);
+    const roomsJoined = await this.roomService.findAllRoomUserNumber(payload.source);
     if (roomsJoined >= 15)
       throw new WsException(
         'You joined too many channels, leave some before creating a new one',
@@ -226,7 +228,7 @@ export class Gateway
       throw new WsException("You're banned from that room");
     if (await this.roomService.isUserInRoom(payload.room, user.uid))
       throw new WsException("You're already in that room");
-    const roomsJoined = await this.roomService.findAllRoomUser(payload.source);
+    const roomsJoined = await this.roomService.findAllRoomUserNumber(payload.source);
     if (roomsJoined >= 15)
       throw new WsException(
         'You joined too many channels, leave some before joining a new one',
@@ -287,14 +289,14 @@ export class Gateway
         return;
       }
     }
-
+    client.join(payload.room);
     this.server.emit(client.id, {
       action: actionTypes.joinRoom,
       target: payload.room,
       hasPassword: hasPassword,
       role: role,
     });
-    this.server.emit(payload.room, {
+    this.server.to(payload.room).emit(payload.room, {
       source: payload.source,
       target: payload.room,
       action: actionTypes.join,
@@ -327,6 +329,8 @@ export class Gateway
     const room = await this.roomService.findOne(payload.room);
     if (!room) return;
     const previousOwner = room.owner.uid;
+    if (await this.roomService.isUserInRoom(payload.room, user.uid) === false)
+      throw new WsException("You're not in that room");
     const owner = await this.roomService.removeUser(room, user.uid);
     const ownerStatus = await this.usersStatusService.getUserStatus(
       owner?.username,
@@ -341,7 +345,8 @@ export class Gateway
         action: actionTypes.owner,
       });
     }
-    this.server.emit(payload.room, {
+    client.leave(payload.room);
+    this.server.to(payload.room).emit(payload.room, {
       source: payload.source,
       target: payload.room,
       action: actionTypes.left,
@@ -402,7 +407,7 @@ export class Gateway
       throw new WsException("You're not an admin of this room");
     if (room.owner.username === payload.target) throw new WsException('You cannot kick the owner');
     await this.roomService.removeUser(room, user.uid);
-    this.server.emit(payload.room, {
+    this.server.to(payload.room).emit(payload.room, {
       source: payload.target,
       target: payload.room,
       action: actionTypes.left,
@@ -411,7 +416,8 @@ export class Gateway
       payload.target,
     );
     if (!targetStatus) return;
-    this.server.emit(targetStatus.clientId, {
+    targetStatus.client.leave(payload.room);
+    this.server.to(targetStatus.clientId).emit(targetStatus.clientId, {
       source: payload.source,
       target: payload.room,
       action: actionTypes.kick,
@@ -456,6 +462,7 @@ export class Gateway
       payload.target,
     );
     if (!targetStatus) return;
+    targetStatus.client.leave(payload.room);
     this.server.emit(targetStatus.clientId, {
       source: payload.source,
       target: payload.room,
@@ -752,9 +759,10 @@ export class Gateway
       )
     )
       throw new WsException('Forbidden');
-    this.invitedChat = this.invitedChat.filter(
-      (obj) => obj.id !== client.id && obj.value !== payload.roomName,
+    const index = this.invitedChat.findIndex(
+      (obj) => obj.id === client.id && obj.value === payload.roomName,
     );
+    this.invitedChat.slice(index, 1);
     const user = await this.userService.findOne(payload.username);
     if (!user) throw new WsException(payload.username + ' not found');
     const userStatus = await this.usersStatusService.getUserStatus(
@@ -762,7 +770,10 @@ export class Gateway
     );
     if (!userStatus || userStatus.clientId !== client.id)
       throw new WsException('Forbidden');
+    if (await this.roomService.isUserInRoom(payload.roomName, user.uid))
+      throw new WsException('You\'re already in that room');
     await this.roomService.addUser(payload.roomName, user, true);
+    userStatus.client.join(payload.roomName);
     this.server.emit(payload.roomName, {
       source: payload.username,
       target: payload.roomName,
@@ -832,10 +843,11 @@ export class Gateway
         (obj) => obj.id === client.id && obj.value === payload.target,
       )
     )
-      throw new WsException('Forbidden');
-    this.askFriend = this.askFriend.filter(
-      (obj) => obj.id !== client.id && obj.value !== payload.target,
+      throw new WsException(payload.target + ' didn\'t ask you to be friend');
+    const index = this.askFriend.findIndex(
+      (obj) => obj.id === client.id && obj.value === payload.target,
     );
+    this.askFriend.slice(index, 1);
     const sourceUser = await this.userService.findOne(payload.source);
     if (!sourceUser) throw new WsException(payload.source + ' not found');
     const targetUser = await this.userService.findOne(payload.target);
@@ -936,16 +948,17 @@ export class Gateway
     const cUserStatus = await this.usersStatusService.getUserStatusByClientId(
       client.id,
     );
+    if (!cUserStatus) throw new WsException('Forbidden');
     if (
       !this.invitedPong.find(
         (obj) => obj.id === client.id && obj.value === payload,
       )
-    )
-      throw new WsException('Forbidden');
+    ) {
+        throw  new WsException(payload + ' is already playing');
+    }
     this.invitedPong = this.invitedPong.filter(
       (obj) => obj.id !== client.id && obj.value !== payload,
     );
-    if (!cUserStatus) throw new WsException('Forbidden');
     const user = await this.userService.findOne(cUserStatus.username);
     if (!user) throw new WsException(cUserStatus.username + ' not found');
     const opponentStatus = await this.usersStatusService.getUserStatus(payload);
@@ -1269,6 +1282,8 @@ export class Gateway
     if (!user) throw new WsException(userStatus + ' not found');
     await this.userService.changeUsername(user, payload);
     await this.usersStatusService.changeUsername(userStatus.username, payload);
+    client.leave(oldUsername + '⌲');
+    client.join(payload + '⌲');
     this.server.emit('newUsername', {
       source: oldUsername,
       target: payload,
@@ -1314,8 +1329,11 @@ export class Gateway
     const userStatusTmp = await this.usersStatusService.getUserStatusByClientId(
       client.id,
       );
-      if (!userStatusTmp) return;
-      console.log('client disconnected: ', client.id, userStatusTmp.username);
+    if (!userStatusTmp) return;
+    console.log('client disconnected: ', client.id, userStatusTmp.username);
+    client.leave(client.id);
+    client.leave(client.id + "GETUID");
+    client.leave(userStatusTmp.username + '⌲');
     const user = await this.userService.findOne(userStatusTmp.username);
     if (userStatusTmp.status === userStatus.playing) {
       const gameRoom = this.gameService.getGameRoom(userStatusTmp.gameRoomId);
@@ -1401,6 +1419,13 @@ export class Gateway
       user.username,
       userStatus.online,
     );
+    const rooms = await this.roomService.findAllRoomUser(user.username);
+    rooms.forEach(element => {
+      client.join(element.roomName);
+    });
+    client.join(client.id);
+    client.join(client.id + "GETUID");
+    client.join(user.username + '⌲');
     console.log('client connected: ', client.id, user.username);
   }
 }
